Update tests according to new permission system

+++ app/Core/Http/Controllers/DiscussionController.php
@@ -44,6 +44,7 @@ class DiscussionController extends Controller
 
     public function delete(Discussion $discussion)
     {
+        $this->authorize('delete', $discussion);
         $discussion->delete();
 
         return response()->json([
