+++ app/Core/Http/Controllers/TaskTagController.php
@@ -10,6 +10,7 @@ class TaskTagController extends Controller
 {
     public function store(Task $task)
     {
+        $this->authorize('attach', Tag::class);
         $task->tags()->attach(request('labels'));
 
         return response()->json([
@@ -20,6 +21,7 @@ class TaskTagController extends Controller
 
     public function delete(Task $task, Tag $tag)
     {
+        $this->authorize('detach', $tag);
         $task->tags()->detach($tag);
 
         return response()->json([
